Add GitHub Actions CI for Linux and Windows builds (#1)
* Add GitHub Actions CI: Linux GCC + Windows MSVC builds

* CI: trigger only on master branch

* CI: fix Qt packages for Linux and Windows jobs

* CI: add Qt5Compat dev package for Linux, downgrade to install-qt-action@v3 for Windows

* CI: remove Windows job, keep Linux only

* Fix GCC warnings: deprecated Qt6 APIs and NULL-in-arithmetic

* Untrack ci_log.txt

* gitignore: exclude ci_log.txt

+++ src/editor/BuildingEditor/buildingtemplates.cpp
@@ -677,7 +677,7 @@ QStringList BuildingTemplate::mTileNames;
 BuildingTemplate::BuildingTemplate() :
     mTiles(TileCount)
 {
-    Q_ASSERT(TileCount == Building::TileCount);
+    Q_ASSERT(static_cast<int>(TileCount) == static_cast<int>(Building::TileCount));
 }
 
 BuildingTemplate::BuildingTemplate(BuildingTemplate *other)
